Editing the while loop to accommodate for initials sClass removal
Issue, the count when it is expected to be 2 is actually 3 therefore
the correct negations need to be done consistently. Also need to
account for when “fields” in a sClass are created the variable that is
returned is a list and so that list needs to be traversed and values
set to null.